+++ .gitignore
@@ -4,6 +4,7 @@ data/
 src/backend/docker-compose-local.yml
 miniconda.sh
 es_data
+node_modules
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
