+++ README.md
@@ -239,3 +239,21 @@ Run Django server:
 ```bash
 (cruise-literature)$ python manage.py runserver YOUR_IP:YOUR_PORT
 ```
+
+
+### 3. Running forntend
+
+1. NPM instalation
+   Make sure you have npm installed!
+   - WSL: https://stackoverflow.com/questions/49919063/installing-npm-node-on-bash-on-ubuntu-on-windows-wsl-what-architecture-does-u
+2. Dependency installation
+```bash
+cd src/frontendApp/
+npm ci
+```
+
+1. Running
+```
+cd src/frontendApp/
+npm start
+```
