--- frontendApp/src/App.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from './components/home/Home';
-import Faq from './components/faq/Faq';
-import './index.css';
-
-function App() {
-  return (
-    <div className="App">
-      <Router>
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/faq" element={<Faq />} />
-        </Routes>
-      </Router>
-    </div>
-  );
-}
-
-export default App;
