+++ src/frontend/public/cruise-logo-small.png
