+++ src/frontend/public/modal.js
