+++ src/frontend/public/sort-paginate-table.js
