+++ src/frontend/src/components/faq/Faq.js
